+++ CHANGELOG.md
@@ -20,11 +20,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Minimum Rhino version required increased to 8.18
 - Renamed `gh_retrieve_components` by `gh_get_available_components`
 - Prevent `GHDocumentUtils.GetObjectsDetails` from generating humanReadable field if value is already human readable (numbers and strings)
 - Renamed `evaluateList` and `filterList` AI tools to `list_evaluate` and `list_filter`
 - Renamed `evaluateText` and `generateText` AI tools to `text_evaluate` and `text_generate`
-- Minimum Rhino version required increased to 8.18
+- Migrated `GhPutTools` to `Utils` in `Core.Grasshopper`
 
 ## [0.3.1-alpha] - 2025-05-06
